Adding in a 'Connection' API in order to start implementing publisherUpdate() and getBusInfo().  Connection state information is necessary for tracking the publisherUpdates() correctly.
New Connection interface and TcpRosConnection implementation provides the logic
for connecting a TCPROS socket.  Migrating TCPROS-specific information out of
the internal.Subscriber, but this is still in progress.

New TopicConnectionInfo data structure for storing actual connection state.

Migrated more of the internal APIs to use *Identifier instances instead of Strings
to make the above easier.

Lots of TODOs.