Beginning cleanup of high-level context and namespace APIs.
The NodeContext API and loading is starting to take shape, though I have
not yet pushed the API into usage.  The next step will be making sure
that everything within the Node instance now correctly uses the NodeContext
to pull in configuration values.

 - New CommandLineLoader implements RosLoader and initializes a NodeContext
 - RosName is now an internal class
 - Test cases for CommandLineLoader (about 80% complete)
 - New CommandLine and EnvironmentVariable interfaces to provide constants for config
 - Context -> NodeContext (less generic name)
 - RosResolver -> NameResolver (Ros prefix wasn't working for me)
 - RosNamespace -> Namespace

For the removal of the 'Ros' prefix, I'm open to other final names, but having
two classes with the Ros prefix just wasn't working for me in practice.