Fix for reporting 0/ bind addresses publicly.  The issue was a bind address being passed into the ProtocolDescription instance.
In order to deal with this, I've taken a different approach than
my previous one with the NodeSocketAddress.  That approach didn't
work because it didn't separate the semantics of binding
addresses from other addresses, and it was really only necessary
for binding addresses.  Thus, I've moved many APIs back to use
InetSocketAddress (no more SocketAddress either) and tried to be
more careful with the use of NodeBindAddress.  NodeBindAddress
is not an extension of SocketAddress anymore because that makes
it too easy to pass into the incorrect APIs.